+++ NEWS.md
@@ -3,6 +3,7 @@
 - Added phrases to extract POS sequences more easily like noun phrases, verb phrases or any sequence of parts of speech tags and their corresponding words
 - Fix issue in txt_nextgram if n was larger than the number of elements in x
 - Fix heap-use-after-free address sanitiser issue
+- Fix runtime error: null pointer passed as argument 1, which is declared to never be null (e.g. udpipe.cpp: 3338)
 
 
 # CHANGES IN udpipe VERSION 0.2